refactor!: retain only the C/C++ module API 🔥
