+++ src/cjs-shim.ts
@@ -2,21 +2,41 @@ import type { ResolvedConfig, Plugin } from 'vite'
 
 export default function cjsShim(): Plugin {
   let config: ResolvedConfig
+  let isCjs: boolean
 
   return {
     name: 'vite-plugin-electron-renderer:cjs-shim',
     apply: 'build',
     configResolved(_config) {
       config = _config
+
+      const output = config.build.rollupOptions.output
+      if (output) {
+        const formats = ['cjs', 'commonjs']
+
+        // https://github.com/electron-vite/vite-plugin-electron/issues/6
+        // https://github.com/electron-vite/vite-plugin-electron/commit/e6decf42164bc1e3801e27984322c41b9c2724b7#r75138942
+        if (
+          Array.isArray(output)
+            // Once an `output.format` is CJS, it is considered as CommonJs
+            ? output.find(o => formats.includes(o.format as string))
+            : formats.includes(output.format as string)
+        ) {
+          isCjs = true
+        }
+      }
     },
     transformIndexHtml(html) {
-      const headRE = /(<\s*?head\s*?>)/
+      if (!isCjs) return
 
-      // ---------------------------------------- shim-require-id
+      const headRE = /(<\s*?head\s*?>)/
       const assetsDir = config.build.assetsDir
+
       if (assetsDir) {
+        // ---------------------------------------- shim-require-id
         const requireIdShim = `<script id="shim-require-id">
 ; (function () {
+  if (typeof require !== 'function') return;
   var Module = require('module');
   var _resolveFilename = Module._resolveFilename;
   Module._resolveFilename = function (request, parent, isMain, options) {
@@ -37,23 +57,9 @@ export default function cjsShim(): Plugin {
       }
 
       // ---------------------------------------- shim-exports
-      const output = config.build.rollupOptions.output
-      if (output) {
-        const formats = ['cjs', 'commonjs']
-
-        // https://github.com/electron-vite/vite-plugin-electron/issues/6
-        // https://github.com/electron-vite/vite-plugin-electron/commit/e6decf42164bc1e3801e27984322c41b9c2724b7#r75138942
-        if (
-          Array.isArray(output)
-            // Once an `output.format` is CJS, it is considered as CommonJs
-            ? output.find(o => formats.includes(o.format as string))
-            : formats.includes(output.format as string)
-        ) {
-          // fix(🐞): `exports is not defined` in "use strict"
-          const exportsShim = `<script id="shim-exports">var exports = typeof module !== 'undefined' ? module.exports : {};</script>`
-          html = html.replace(headRE, '$1\n' + exportsShim)
-        }
-      }
+      // fix(🐞): `exports is not defined` in "use strict"
+      const exportsShim = `<script id="shim-exports">var exports = typeof module !== 'undefined' ? module.exports : {};</script>`
+      html = html.replace(headRE, '$1\n' + exportsShim)
 
       return html
     }
